backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Geospatial dependencies (with NumPy 1.x compatibility)
numpy<2.0.0
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
geoalchemy2==0.14.2

# Pydantic and validation
pydantic[email]==2.5.0
pydantic-core==2.14.1
email-validator==2.2.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utility packages
python-dotenv==1.0.0
requests==2.31.0

# Database connection and SSL support
certifi==2023.11.17

# Optional: Development and testing (uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
